remove unused code
most of this removed code is related to other oauth provders. This will
be slowly brought back.
