--- auth/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "auth"
-version = "0.1.0"
-edition = "2021"
-description = "Auth backend"
-
-[lib]
-crate-type = ["cdylib"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ft-sdk = { path = "../../ft-sdk/ft-sdk/", features = ["sqlite-default", "auth-provider"] }
-serde = {  version = "1", features = ["derive"] }
-serde_json = "1"
-http = "1"
-bytes = "1"
-diesel = "2"
-thiserror = "1"
-argon2 = { version = "0.5" }
-chrono = { version = "0.4", default-features = false, features = ["serde"] }
-include_dir = "0.7"
-validator = "0.18"
