--- auth/rust-toolchain.toml
@@ -1,3 +1,0 @@
-[toolchain]
-channel = "1.78.0"
-
