--- auth/src/email_password/email_confirmation_sent.rs
@@ -1,20 +1,0 @@
-use crate::auth::email_password::email_confirmation_sent_ftd;
-
-// TODO: this is unused right now
-// the user is immediately logged in after creating an account and this page is never
-// visited/redirected to
-pub(crate) async fn email_confirmation_sent(
-    req_config: &mut fastn_core::RequestConfig,
-) -> fastn_core::Result<fastn_core::http::Response> {
-    let main = fastn_core::Document {
-        package_name: req_config.config.package.name.clone(),
-        id: fastn_core::auth::Route::EmailConfirmationSent.to_string(),
-        content: email_confirmation_sent_ftd().to_string(),
-        parent_path: fastn_ds::Path::new("/"),
-    };
-
-    let resp =
-        fastn_core::package::package_doc::read_ftd(req_config, &main, "/", false, false).await?;
-
-    Ok(resp.into())
-}
