--- auth/src/handlers/mod.rs
@@ -1,5 +1,0 @@
-mod create_account;
-mod login;
-
-pub use create_account::CreateAccount;
-pub use login::Login;
