+++ ft-email-auth-provider/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "ft-email-auth-provider"
+name = "ft_email_auth_provider"
 version = "0.1.0"
 edition = "2021"
-description = "Auth backend"
+description = "Email auth provider"
 
 [lib]
 crate-type = ["cdylib"]
@@ -10,7 +10,7 @@ crate-type = ["cdylib"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ft-sdk = { path = "../../ft-sdk/ft-sdk", features = ["auth-providers"] }
+ft-sdk = { path = "../../ft-sdk/ft-sdk/", features = ["sqlite-default", "auth-provider"] }
 serde = {  version = "1", features = ["derive"] }
 serde_json = "1"
 http = "1"
@@ -19,3 +19,5 @@ diesel = "2"
 thiserror = "1"
 argon2 = { version = "0.5" }
 chrono = { version = "0.4", default-features = false, features = ["serde"] }
+include_dir = "0.7"
+validator = "0.18"
