+++ ft-email-auth-provider/src/lib.rs
@@ -8,8 +8,6 @@ mod utils;
 
 pub const PROVIDER_ID: &str = "email";
 
-// TODO: login
-// TODO: create-account
 // TODO: logout
 
 #[no_mangle]
